moved get_node_outputs_changes and typing changes
